refactor: 컴포넌트 내 폴더 구조 통일, 중복 타이틀 제거
