+++ app/(anon)/@detail/steps/[step-number]/[detail]/_components/contents/guaranteeLimit/guaranteeLimitContainer/GuaranteeLimitContainer.styles.ts
