+++ app/(anon)/@detail/steps/[step-number]/[detail]/_components/contents/guaranteeLimit/guaranteeLimitOutput/GuaranteeLimitOutput.styles.ts
